+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `RevitVersionMinimal` `PropertyGroup` to set minimal version for `XXX_OR_GREATER`. (Default: `2019`)
 - Update `XXX_OR_GREATER` implementation based on [discussions](https://github.com/Nice3point/RevitTemplates/discussions/43).
 - `Test` project with `CountTests` for version `2017`, `2019`, `2021` and `2025`. (`RevitVersionMinimal` to `2017`) 
+- Change to `AddImplicitDefineConstants` and `DefineConstants`.
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0